# Bool writer optimization