This is slightly faster due to the changes in `put` code, other changes are just comments. Changes in `finish` do not change behavior as it can be checked simply that putting 32 false with dummy branches into the stream just push out 32 bits of current `value` irrespective of initial `range`.

main 82d6547e924b3b54af57c576438fa63b94eddc2c
```
2024-11-15T21:03:34.699Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-15T21:03:35.173Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-15T21:03:36.610Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 9632ms of CPU time in 1436ms of wall time
2024-11-15T21:03:36.610Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-15T21:03:38.216Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324076 bytes (compression = 28.0%)
2024-11-15T21:03:38.216Z INFO  [lepton_jpeg_util] Main thread CPU: 3517ms, Worker thread CPU: 20710 ms, walltime: 3517 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       871 003 064      cache-references                                                        (41,51%)
        86 439 870      cache-misses                     #    9,92% of all cache refs           (41,52%)
    15 795 148 123      cycles                                                                  (41,72%)
       834 608 725      ic_fetch_stall.ic_stall_back_pressure                                        (41,87%)
     1 014 898 692      stalled-cycles-frontend          #    6,43% frontend cycles idle        (42,14%)
    39 107 914 014      instructions                     #    2,48  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (42,43%)
     4 396 207 542      branch-instructions                                                     (42,50%)
       162 547 993      branch-misses                    #    3,70% of all branches             (42,32%)
     5 434 637 214      ic_fetch_stall.ic_stall_any                                             (42,14%)
        45 374 631      ic_fetch_stall.ic_stall_dq_empty                                        (41,99%)
        72 770 962      l2_cache_misses_from_ic_miss                                            (41,75%)
     2 149 017 778      l2_latency.l2_cycles_waiting_on_fills                                        (41,47%)
           184 005      faults                                                                
                 1      migrations                                                            

       3,551824055 seconds time elapsed

       3,224073000 seconds user
       0,324906000 seconds sys
```

this PR
```
2024-11-16T11:33:06.593Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-16T11:33:07.076Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-16T11:33:08.477Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 9262ms of CPU time in 1400ms of wall time
2024-11-16T11:33:08.477Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-16T11:33:10.082Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324076 bytes (compression = 28.0%)
2024-11-16T11:33:10.082Z INFO  [lepton_jpeg_util] Main thread CPU: 3488ms, Worker thread CPU: 20322 ms, walltime: 3488 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       853 708 322      cache-references                                                        (41,67%)
        73 780 380      cache-misses                     #    8,64% of all cache refs           (41,99%)
    15 675 228 373      cycles                                                                  (42,20%)
       845 297 301      ic_fetch_stall.ic_stall_back_pressure                                        (42,21%)
     1 078 381 569      stalled-cycles-frontend          #    6,88% frontend cycles idle        (42,22%)
    38 226 009 631      instructions                     #    2,44  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (42,18%)
     4 363 626 024      branch-instructions                                                     (42,06%)
       164 603 035      branch-misses                    #    3,77% of all branches             (41,86%)
     5 214 932 302      ic_fetch_stall.ic_stall_any                                             (41,71%)
        40 234 830      ic_fetch_stall.ic_stall_dq_empty                                        (41,71%)
        60 213 231      l2_cache_misses_from_ic_miss                                            (41,75%)
     2 112 292 962      l2_latency.l2_cycles_waiting_on_fills                                        (41,64%)
           183 960      faults                                                                
                 1      migrations                                                            

       3,523920131 seconds time elapsed

       3,190726000 seconds user
       0,328662000 seconds sys
```